refactor(tui): return screen view-model ownership to each screen module (#434)
`view_model.rs` was a shallow module: 1,175 lines declaring 31 public types,
but `build_view_model` is six lines dispatching through
`screens::lookup(...).build_vm`, so every per-screen builder already lived in
its own screen module. Only `ViewModel`, `ChromeVm`, and `ScreenVm` were
cross-screen framework; the rest was single-screen presentation data whose
builder, paint function, and tests were all somewhere else. Two thirds of the
file was a test block where 21 of 22 tests reached one screen's view model by
calling `build_view_model` and destructuring `.screen`.

Each screen module now owns the view-model types it owns, with their tests.
The stated rule: can one screen claim this type? If yes it belongs to that
screen; if no — several screens use it as equals to describe themselves — it
belongs to the shared renderer it is the contract for. Caller count is only a
proxy, and `DiffVm` is where the proxy fails: two screen callers, but it *is*
the Diff screen and Confirm is a named borrower.

- `ScreenVm` joins `screens::lookup`, whose `build_vm` column constructs its
  variants, so the screen list exists once rather than twice.
- `ListPaneVm` / `ListPaneEmpty` / `RowVm` / `RowEmphasis` join
  `render/list_pane.rs`, the renderer they are the input contract for.
- `PaneTitleVm` goes one step further to `render/text_fit.rs`: `ListPaneVm`
  only holds one and hands it on, while `fit_title` is the only code that
  splits it into segments.
- `file_ext` and `gist_row_display` are not view models and join
  `render/labels.rs`; `build_background_screen_vm` joins Palette, whose policy
  its `None` arms encode.
- Every moved declaration drops from `pub` to `pub(crate)`, matching the
  destination modules and no longer relying on a private module as cover. No
  facade re-export: `src/tui/mod.rs`'s old `pub(crate) use view_model::{...}`
  is gone and call sites name the real module.

`view_model.rs` is 88 lines: `ViewModel`, `ChromeVm`, the two functions that
build them, and two smoke tests that still exercise the whole path.

Behaviour-preserving. 843 tests before and after; `assert!`/`assert_eq!`/
`assert_ne!` counts are unchanged at 461/682/2. The 27 fewer `panic!`s are the
`other => panic!("expected X")` arms of matches that became direct builder
calls, where the type is now guaranteed by the compiler.
